requirements.txt
streamlit
sentence-transformers
scikit-learn
python-docx
PyPDF2
tf-keras
st_aggrid
Pillow

streamlit
sentence-transformers
PyPDF2
python-docx
scikit-learn
numpy
pandas
plotly
openai
spacy
nltk
transformers
torch